loan_approval_classification/requirements.txt
pandas
ydata-profiling
scikit-learn
lightgbm
optuna
pillow
requests
streamlit